add eslint and refactor code
