+++ pokedex/.prettierrc
@@ -1,7 +1,7 @@
 {
   "trailingComma": "es5",
   "bracketSpacing": true,
-  "tabWidth": 4,
+  "tabWidth": 2,
   "semi": true,
   "singleQuote": true,
   "printWidth": 80
