+++ pokedex/package.json
@@ -17,8 +17,16 @@
     "sharp": "^0.32.1"
   },
   "devDependencies": {
+    "@next/eslint-plugin-next": "^13.4.2",
     "@types/react": "^18.0.33",
     "autoprefixer": "^10.4.14",
+    "eslint": "^8.40.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "postcss": "^8.4.21",
     "tailwindcss": "^3.3.1"
   }
