--- pokedex/src/components/Container.js
@@ -1,12 +1,0 @@
-import Header from './Header';
-
-function Container({ children }) {
-    return (
-        <div className="bg-primary min-h-screen">
-            <Header />
-            <main>{children}</main>
-        </div>
-    );
-}
-
-export default Container;
