+++ pokedex/src/components/LoadingScreen.jsx
@@ -1,0 +1,12 @@
+import Lottie from 'lottie-react';
+import animationData from '/public/animations/pokeball-loading.json';
+
+function LoadingScreen() {
+  return (
+    <div className="pointer-events-none absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2">
+      <Lottie animationData={animationData} loop={true} />
+    </div>
+  );
+}
+
+export default LoadingScreen;
