+++ pokedex/src/context/GlobalContext.js
@@ -1,3 +1,4 @@
 import { createContext } from 'react';
 
-export const GlobalContext = createContext();
+const GlobalContext = createContext();
+export default GlobalContext;
