+++ pokedex/src/hook/useGlobalContext.js
@@ -1,14 +1,14 @@
-import { GlobalContext } from '@/context/GlobalContext';
 import { useContext } from 'react';
+import GlobalContext from '@/context/GlobalContext';
 
 const useGlobalContext = () => {
-    const context = useContext(GlobalContext);
+  const context = useContext(GlobalContext);
 
-    if (context === undefined) {
-        throw new Error('Não está dentro do contexto');
-    }
+  if (context === undefined) {
+    throw new Error('Não está dentro do contexto');
+  }
 
-    return context;
+  return context;
 };
 
 export default useGlobalContext;
