+++ pokedex/src/pages/_app.jsx
@@ -3,11 +3,11 @@ import Container from '@/components/Container';
 import GlobalProvider from '@/context/GlobalProvider';
 
 export default function App({ Component, pageProps }) {
-    return (
-        <GlobalProvider>
-            <Container>
-                <Component {...pageProps} />
-            </Container>
-        </GlobalProvider>
-    );
+  return (
+    <GlobalProvider>
+      <Container>
+        <Component {...pageProps} />
+      </Container>
+    </GlobalProvider>
+  );
 }
